pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = ""
authors = ["Michail Fedyaev <drovosekov1910@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
